refactor: rename Description-styling file
